[ensurepip patch] Add '--pre' to args only when installing
As it stands right now _run_pip always adds the --pre option,
however only "install", "list" and "wheel" actually support it,
which makes the python test suite fail when it tries to use "uninstall".

I only check for "install" and not for "list" or "wheel" because
"install" is the only one from that list that is used by this module.